pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devmate"
version = "0.3.0"
description = "The friendly internal developer platform CLI."
dependencies = [
    "typer[all]",
    "rich",  
    "docker",
    "requests",
    "GitPython"
]
requires-python = ">=3.10"

[project.scripts]
devmate = "app.main:app"
devctl = "app.main:app"
mate = "app.main:app"